Added MSSQL and other SQL scripts
There's no real organization to this right
now but I know there should be. Most queries
are ran on demand and are simple enough to not
want to create a separate file for. I'll work on
sharing common queries and SQL views/dashboards
which take advantage of all the data available.